cryptoengine/services/market-data/requirements.txt
ccxt>=4.2.0
aiohttp>=3.9.0
websockets>=12.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.0
TA-Lib>=0.4.28
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
pydantic>=2.6.0
structlog>=24.1.0
pyyaml>=6.0.0
orjson>=3.9.0
